rename SpecificMethodIntenralTypeRector to AssertTrueFalseInternalTypeToSpecificMethodRector
