+++ src/config/level/phpunit/phpunit-specific-method.yml
@@ -1,6 +1,6 @@
 rectors:
     Rector\Rector\Contrib\PHPUnit\SpecificMethod\AssertTrueFalseToSpecificMethodRector: ~
     Rector\Rector\Contrib\PHPUnit\SpecificMethod\AssertSameBoolNullToSpecificMethodRector: ~
+    Rector\Rector\Contrib\PHPUnit\SpecificMethod\AssertTrueFalseInternalTypeToSpecificMethodRector: ~
     Rector\Rector\Contrib\PHPUnit\SpecificMethodCountRector: ~
-    Rector\Rector\Contrib\PHPUnit\SpecificMethodInternalTypeRector: ~
     Rector\Rector\Contrib\PHPUnit\SpecificMethodObjectAttributeRector: ~
